fix: use maps for alarm overwrites to correctly target individual alarms (#72)
## Description
<!--- Describe your changes in detail -->
This changes the alarm inputs to map of objects. They will be merged
during execution with a set of default defined per individual alarm to
allow configuring each alarm separately.

## Motivation and Context
When using the gosoline monitoring provider in combination with the
three alarm-{consumer|gateway|kinsumer}, this creates individual alarms
based on the application configuration (e.g. different success rate
alarms for three different kinsumers).
The previous alarm overwrites passed through the variables didn't allow
to target the individual alarms; instead they were applied to all of
them.

## Breaking Changes
When you're passing any of the alarm* variables, this is a breaking
change. The structure change was required to target individual alarms
instead of applying the change to all alarms.

## How Has This Been Tested?
I verified the results with an instance of each alarm type.